preprocessing/pyproject.toml
[project]
name = "preprocessing"
version = "0.1.0"
description = "Preprocessing service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0,<4.0.0",
    "fastapi>=0.104.1",
    "firebase-admin>=6.0.0",
    "google-cloud-storage==2.10.0",
    "google-cloud-firestore>=2.11.0",
    "pandas>=2.1.3",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "uvicorn[standard]>=0.24.0",
    "werkzeug>=3.0.1",
    "transformers>=4.56.0",
    "torch>=2.0.0",
    "sentencepiece",
    "nltk>=3.8",
    "sacremoses>=0.1.1",
    "google-genai>=1.4.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=20.0.0",
    "pillow>=10.0.0",
    "pymupdf>=1.26.4",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "beautifulsoup4>=4.14.2",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
